LayoutTests/fast/files/blob-constructor-expected.txt

Test the Blob constructor.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS (new Blob()) instanceof window.Blob is true
PASS (new Blob([])) instanceof window.Blob is true
PASS (new Blob(['hello'])) instanceof window.Blob is true
PASS (new Blob(['hello'], {})) instanceof window.Blob is true
PASS (new Blob(['hello'], {type:'text/html'})) instanceof window.Blob is true
PASS (new Blob(['hello'], {type:'text/html', endings:'native'})) instanceof window.Blob is true
PASS (new Blob(['hello'], {type:'text/html', endings:'transparent'})) instanceof window.Blob is true
PASS new Blob('hello') threw exception TypeError: First argument of the constructor is not of type Array.
PASS new Blob(0) threw exception TypeError: First argument of the constructor is not of type Array.
PASS (new Blob([])) instanceof window.Blob is true
PASS (new Blob([new ArrayBuffer(8)])) instanceof window.Blob is true
PASS (new Blob(['stringPrimitive'])) instanceof window.Blob is true
PASS (new Blob([String('stringObject')])) instanceof window.Blob is true
PASS (new Blob([new Blob])) instanceof window.Blob is true
PASS (new Blob([new Blob([new Blob])])) instanceof window.Blob is true
PASS (new Blob([12])).size is 2
PASS (new Blob([[]])).size is 0
PASS (new Blob([{}])).size is 15
PASS (new Blob([document])).size is 21
PASS (new Blob([toStringingObj])).size is 8
PASS new Blob([throwingObj]) threw exception error.
PASS (new Blob([], {unknownKey:'value'})) instanceof window.Blob is true
PASS new Blob([], {endings:'illegalValue'}) threw exception Error: INVALID_STATE_ERR: DOM Exception 11.
PASS new Blob([], {endings:throwingObj}) threw exception error.
PASS new Blob([], {type:throwingObj}) threw exception error.
PASS (new Blob([], null)) instanceof window.Blob is true
PASS (new Blob([], undefined)) instanceof window.Blob is true
PASS (new Blob([], 123)) instanceof window.Blob is true
PASS (new Blob([], 123.4)) instanceof window.Blob is true
PASS (new Blob([], [])) instanceof window.Blob is true
PASS (new Blob([], true)) instanceof window.Blob is true
PASS (new Blob([], 'abc')) instanceof window.Blob is true
PASS (new Blob([], /abc/)) instanceof window.Blob is true
PASS (new Blob([], function () {})) instanceof window.Blob is true
PASS (new Blob([], {type:'text/html'})).type is 'text/html'
PASS (new Blob([], {type:'text/html'})).size is 0
PASS window.Blob.length is 2
PASS successfullyParsed is true

TEST COMPLETE
